Updated: Installer build system, Updated script version 2.2
Here we have passed the large update to the build system and mechanics
on how cyberfox installers are generated.

This new process that is now 99% automated should reduce packaging time
overall.

How it works is the browser files are placed in there folders and the
app.config gets updated with the required parameters
for (architecture), (package) & (version), This change removes having to
maintain 4 individual scripts as they are now one
script that does all.

We will add the beta installer to this new system once we decide if we
are going to keep it full time.